Python: reorganisation of the full configuration directory
