+++ Cards/lpair_ttbar_cfg.py
@@ -1,7 +1,7 @@
 import Config.Core as cepgen
 from Config.PDG_cfi import PDG
-from Config.generator_cff import generator as _gen
-#from Config.Timer_cfi import timer  # enable timing framework
+from Config.generator_cfi import generator as _gen
+#from Config.timer_cfi import timer  # enable timing framework
 
 process = cepgen.Module('lpair',
     processParameters = cepgen.Parameters(
