+++ python/Config/Core.py
@@ -5,11 +5,11 @@ A collection of tools for Python steering cards definition
 '''
 
 #--- core components includes
-from .containers_cfi import Module, Parameters, Sequence
-from .Logger_cfi import Logging
+from .containers_cff import Module, Parameters, Sequence
+from .logger_cfi import Logging
 
 #--- physics-level includes
-from .StructureFunctions_cfi import StructureFunctions
-from .SigmaRatio_cfi import SigmaRatio
-from .ProcessMode_cfi import ProcessMode
+from .StructureFunctions_cff import StructureFunctions
+from .SigmaRatio_cff import SigmaRatio
+from .ProcessMode_cff import ProcessMode
