--- python/Config/Hadronisation/photos_cff.py
@@ -1,4 +1,0 @@
-from Config.containers_cfi import Module
-
-photos = Module('photos',
-)
