+++ python/Config/Hadronisation/photos_cfi.py
@@ -1,0 +1,4 @@
+from Config.containers_cff import Module
+
+photos = Module('photos',
+)
