+++ python/Config/Hadronisation/pythia8_cfi.py
@@ -1,7 +1,7 @@
-from Config.containers_cfi import Module
+from Config.containers_cff import Module
 
-from Config.Hadronisation.pythia8Defaults_cfi import pythia8Defaults
-from Config.Hadronisation.pythia8Tunes_cfi import pythia8CUEP8M1Settings
+from Config.Hadronisation.pythia8Defaults_cff import pythia8Defaults
+from Config.Hadronisation.pythia8Tunes_cff import pythia8CUEP8M1Settings
 
 pythia8 = Module('pythia8',
     seed = 1000,
