--- python/Config/Integration/foam_cff.py
@@ -1,5 +1,0 @@
-from Config.containers_cfi import Module
-
-integrator = Module('Foam',
-    verbose = 0,
-)
