+++ python/Config/Integration/foam_cfi.py
@@ -1,0 +1,5 @@
+from Config.containers_cff import Module
+
+foam = Module('Foam',
+    verbose = 0,
+)
