+++ python/Config/Integration/plain_cfi.py
@@ -1,0 +1,7 @@
+from Config.Gsl_cff import GslRngEngine
+from Config.containers_cff import Module
+
+plain = Module('plain',
+    numFunctionCalls = 1000000,
+    rngEngine = GslRngEngine.MT19937,
+)
