+++ python/Config/OutputModule/rootTree_cfi.py
@@ -1,4 +1,4 @@
-from Config.containers_cfi import Module
+from Config.containers_cff import Module
 
 rootTree = Module('root_tree')
 rootTreeCompressed = Module('root_tree',
