--- python/Config/Timer_cfi.py
@@ -1,3 +1,0 @@
-from .containers_cfi import Parameters
-
-timer = Parameters()
