+++ python/Config/timer_cfi.py
@@ -1,0 +1,3 @@
+from .containers_cff import Parameters
+
+timer = Parameters()
